Add files via upload
This is the script I used to train and test the different classification algorithms for the BIGMaC project. This script is heavily derived from the resources cited in the file, and I recommend using those instead for application in your work, but I am also uploading this script for educational and transparency purposes.